# db: apply TrySeekUsingNext to level metadata