--- a/testdata/level_iter_seek
+++ b/testdata/level_iter_seek
@@ -306,11 +306,11 @@ j.SET.7:j
 
 # Seeks to immediately following keys.
 iter
-seek-prefix-ge a true
+seek-prefix-ge a false
 seek-prefix-ge a true
 seek-prefix-ge b true
 next
-seek-prefix-ge c true
+seek-prefix-ge c false
 seek-prefix-ge d true
 seek-prefix-ge f true
 seek-prefix-ge g true
@@ -332,7 +332,7 @@ j/<invalid>#7,1:j
 
 # Seeks to keys that are in the next file, so cannot use Next.
 iter
-seek-prefix-ge a true
+seek-prefix-ge a false
 seek-prefix-ge e true
 seek-prefix-ge i true
 seek-prefix-ge j true